requirements.txt
entsoe-py
pandas
streamlit
python-dotenv
plotly
